feat: Implement Logto IAM for authentication
This commit replaces the existing mock authentication system with Logto Identity and Access Management.

The changes include:
- Added the `@logto/next` dependency.
- Configured the Logto SDK with the provided application credentials.
- Created API routes and server actions to handle sign-in, sign-out, and the Logto callback.
- Implemented an `AuthButton` component in the header to provide a way for users to sign in and out.
- Removed the old authentication system, including the login, register, and verify-email pages, the `AuthContext`, and related logic in the main layout.
- Refactored all pages that used the old `AuthContext` to remove the dependency.
