+++ src/app/organizations/new/NewOrganizationForm.tsx
@@ -1,0 +1,112 @@
+'use client';
+
+import { useState, useEffect } from "react";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import { useOrganization } from "@/contexts/OrganizationContext";
+import type { Organization } from "@/contexts/OrganizationContext";
+import { useRouter } from "next/navigation";
+import { useToast } from "@/hooks/use-toast";
+
+export function NewOrganizationForm({ user }: { user: { name: string, sub: string } }) {
+    const { addOrganization, addDefaultDataForNewOrg, organizations, selectOrganization } = useOrganization();
+    const router = useRouter();
+    const { toast } = useToast();
+    const [name, setName] = useState('');
+    const [isCreating, setIsCreating] = useState(false);
+
+    useEffect(() => {
+        // If an existing user with orgs lands here, send them away.
+        if (user && organizations.length > 0) {
+            router.push('/organizations');
+        }
+    }, [user, router, organizations]);
+
+
+    const handleCreate = () => {
+        if (isCreating) return;
+        setIsCreating(true);
+
+        if (!name.trim()) {
+             toast({
+                variant: "destructive",
+                title: "Organization name required",
+                description: "Please enter a name for your organization.",
+            });
+            setIsCreating(false);
+            return;
+        }
+
+        if (organizations.some(org => org.name.toLowerCase() === name.trim().toLowerCase())) {
+            toast({
+                variant: "destructive",
+                title: "Organization name exists",
+                description: "An organization with this name already exists. Please choose a different name.",
+            });
+            setIsCreating(false);
+            return;
+        }
+
+        const newOrg: Organization = {
+            id: `org-${Date.now()}`,
+            name: name.trim(),
+            description: 'Your first organization',
+            type: 'Client',
+            subscribers: 1, // Starts with 1 default user
+            status: 'Trial',
+        };
+
+        // Add the organization but don't auto-select yet. We will do it manually.
+        addOrganization(newOrg, false);
+
+        // This logic runs for the very first organization created by a user.
+        if (user && organizations.length === 0) {
+            addDefaultDataForNewOrg();
+            // Set flags to trigger onboarding hints on the subscribers page.
+            localStorage.setItem(`onboarding_show_groups_hint_${user.name}`, 'true');
+        }
+
+        // Manually select the new org. The callback ensures the redirect
+        // happens only after the state has been successfully updated.
+        selectOrganization(newOrg, () => {
+            router.push('/subscribers');
+        });
+    };
+
+    if (!user) {
+        return (
+             <div className="flex h-screen items-center justify-center text-lg font-semibold">
+                Loading user data...
+            </div>
+        )
+    }
+
+    return (
+        <div className="relative flex h-screen items-center justify-center bg-background px-4">
+           <div className="w-full max-w-xl text-left">
+                <h1 className="text-4xl md:text-5xl font-bold font-headline text-foreground">
+                    Welcome, {user?.name}!
+                </h1>
+                <div className="mt-12 space-y-4">
+                     <Label htmlFor="org-name" className="text-xl text-muted-foreground">
+                        Enter your organization name
+                    </Label>
+                    <Input
+                        id="org-name"
+                        value={name}
+                        onChange={(e) => setName(e.target.value)}
+                        placeholder="e.g. My Company"
+                        className="h-14 text-2xl"
+                    />
+                </div>
+
+                <div className="mt-16 flex justify-end">
+                     <Button onClick={handleCreate} size="lg" disabled={isCreating}>
+                        {isCreating ? 'Creating...' : 'Next >'}
+                    </Button>
+                </div>
+           </div>
+        </div>
+    );
+}
